+++ src/Sulu/Bundle/MediaBundle/Controller/AbstractMediaController.php
@@ -11,6 +11,7 @@
 
 namespace Sulu\Bundle\MediaBundle\Controller;
 
+use Sulu\Bundle\MediaBundle\Media\FormatManager\FormatManagerInterface;
 use Sulu\Bundle\MediaBundle\Media\Manager\MediaManagerInterface;
 use Sulu\Component\Rest\RestController;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
@@ -22,13 +23,23 @@ use Symfony\Component\HttpFoundation\Request;
 class AbstractMediaController extends RestController
 {
     /**
-     * getMediaManager.
+     * Returns media-manager.
      *
      * @return MediaManagerInterface
      */
     protected function getMediaManager()
     {
         return $this->get('sulu_media.media_manager');
+    }
+
+    /**
+     * Returns format-manager.
+     *
+     * @return FormatManagerInterface
+     */
+    protected function getFormatManager()
+    {
+        return $this->get('sulu_media.format_manager');
     }
 
     /**