+++ src/Sulu/Bundle/WebsiteBundle/Routing/PortalLoader.php
@@ -11,7 +11,7 @@
 
 namespace Sulu\Bundle\WebsiteBundle\Routing;
 
-use Sulu\Component\Webspace\Manager\WebspaceManagerInterface;
+use Sulu\Component\Webspace\Analyzer\RequestAnalyzerInterface;
 use Symfony\Component\Config\Loader\Loader;
 use Symfony\Component\Routing\Route;
 use Symfony\Component\Routing\RouteCollection;
@@ -23,53 +23,39 @@ use Symfony\Component\Routing\RouteCollection;
 class PortalLoader extends Loader
 {
     /**
-     * @var WebspaceManagerInterface
-     */
-    private $webspaceManager;
-
-    /**
-     * @var string
-     */
-    private $environment;
-
-    /**
-     * @var RouteCollection
-     */
-    private $collection;
-
-    public function __construct(WebspaceManagerInterface $webspaceManager, $environment)
-    {
-        $this->webspaceManager = $webspaceManager;
-        $this->environment = $environment;
-    }
-
-    /**
      * {@inheritdoc}
      */
     public function load($resource, $type = null)
     {
-        $this->collection = new RouteCollection();
+        $collection = new RouteCollection();
 
         /** @var Route[] $importedRoutes */
         $importedRoutes = $this->import($resource, null);
 
+        $condition = sprintf(
+            'request.get("_sulu").getAttribute("portalInformation").getType() === %s',
+            RequestAnalyzerInterface::MATCH_TYPE_FULL
+        );
+
         foreach ($importedRoutes as $importedRouteName => $importedRoute) {
-            $this->collection->add(
+            $importedCondition = $importedRoute->getCondition();
+
+            $collection->add(
                 $importedRouteName,
                 new PortalRoute(
                     '{prefix}' . $importedRoute->getPath(),
                     $importedRoute->getDefaults(),
                     array_merge(['prefix' => '.*', 'host' => '.+'], $importedRoute->getRequirements()),
                     $importedRoute->getOptions(),
-                    '{host}',
+                    (!empty($importedRoute->getHost()) ? $importedRoute->getHost() : '{host}'),
                     $importedRoute->getSchemes(),
                     $importedRoute->getMethods(),
-                    $importedRoute->getCondition()
+                    $condition . (!empty($importedCondition) ? ' and (' . $importedCondition . ')' : '')
                 )
             );
         }
 
-        return $this->collection;
+        return $collection;
     }
 
     /**