+++ src/Sulu/Component/Rest/ListBuilder/Doctrine/FieldDescriptor/DoctrineFieldDescriptor.php
@@ -12,6 +12,7 @@
 namespace Sulu\Component\Rest\ListBuilder\Doctrine\FieldDescriptor;
 
 use JMS\Serializer\Annotation\ExclusionPolicy;
+use Sulu\Component\Rest\ListBuilder\FieldDescriptorInterface;
 
 /**
  * This class defines the necessary information for a field to resolve it within a Doctrine Query for the ListBuilder.
@@ -56,7 +57,18 @@ class DoctrineFieldDescriptor extends AbstractDoctrineFieldDescriptor
         $editable = false,
         $cssClass = ''
     ) {
-        parent::__construct($name, $translation, $disabled, $default, $type, $width, $minWidth, $sortable, $editable, $cssClass);
+        parent::__construct(
+            $name,
+            $translation,
+            $disabled,
+            $default,
+            $type,
+            $width,
+            $minWidth,
+            $sortable,
+            $editable,
+            $cssClass
+        );
 
         $this->fieldName = $fieldName;
         $this->entityName = $entityName;
@@ -99,5 +111,18 @@ class DoctrineFieldDescriptor extends AbstractDoctrineFieldDescriptor
     public function getJoins()
     {
         return $this->joins;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function compare(FieldDescriptorInterface $other)
+    {
+        if (!$other instanceof self) {
+            return false;
+        }
+
+        return $this->getEntityName() === $other->getEntityName()
+            && $this->getFieldName() === $other->getFieldName();
     }
 }